Pull classes up
