+++ okhttp-clients/src/main/java/com/palantir/conjure/java/okhttp/RemotingOkHttpClient.java
@@ -17,7 +17,6 @@
 package com.palantir.conjure.java.okhttp;
 
 import com.palantir.conjure.java.client.config.ClientConfiguration;
-import com.palantir.conjure.java.client.config.ImmutablesStyle;
 import com.palantir.conjure.java.client.config.NodeSelectionStrategy;
 import com.palantir.logsafe.SafeArg;
 import com.palantir.logsafe.exceptions.SafeIllegalStateException;
@@ -30,7 +29,6 @@ import java.util.function.Supplier;
 import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
-import org.immutables.value.Value;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -120,30 +118,6 @@ final class RemotingOkHttpClient extends ForwardingOkHttpClient {
                 .tag(AttemptSpan.class, AttemptSpan.createAttempt(entireSpan, 0))
                 .tag(SettableDispatcherSpan.class, SettableDispatcherSpan.create())
                 .build();
-    }
-
-    interface EntireSpan extends Supplier<DetachedSpan> {}
-
-    @Value.Immutable
-    @ImmutablesStyle
-    interface AttemptSpan {
-        @Value.Default
-        default int attemptNumber() {
-            return 0;
-        }
-
-        DetachedSpan attemptSpan();
-
-        static AttemptSpan createAttempt(DetachedSpan entireSpan, int attemptNumber) {
-            return ImmutableAttemptSpan.builder()
-                    .attemptNumber(attemptNumber)
-                    .attemptSpan(entireSpan.childDetachedSpan("OkHttp: Attempt " + attemptNumber))
-                    .build();
-        }
-
-        default AttemptSpan nextAttempt(DetachedSpan entireSpan) {
-            return createAttempt(entireSpan, attemptNumber() + 1);
-        }
     }
 
     private HttpUrl getNewRequestUrl(HttpUrl requestUrl) {
